requirements.txt
# Trading Application Dependencies

# Core dependencies
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0

# GUI Framework
PySide6>=6.5.0

# Technical Analysis (optional, can use pandas for EMA/ATR)
# ta-lib  # Uncomment if you want to use TA-Lib instead of pandas

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
